+++ api/src/opentrons/protocols/execution/execute_json_v5.py
@@ -2,7 +2,7 @@ import logging
 from typing import TYPE_CHECKING
 from opentrons.types import Point
 from opentrons.protocols.execution.types import LoadedLabware, Instruments
-from .execute_v3 import _get_well
+from .execute_json_v3 import _get_well
 if TYPE_CHECKING:
     from opentrons_shared_data.protocol.dev_types import (
         MoveToWellParams
